nb: Let signal emission machinery handle NbApplication::create-window
Instead of emitting the signal from the default implementation of
NbApplication.create_window(), tie the virtual function to the signal
emission, and make the default implementation warn if not overriden.
This simplifies code, is more correct, and makes things work in the
usual way for signals which have a corresponding virtual method.